move UploadImage to share folder
